--- client/assets/styles.css
@@ -1,36 +1,0 @@
-html {
-  height: 100%;
-  box-sizing: border-box;
-}
-
-body {
-  position: relative;
-  margin: 0;
-  padding-bottom: 6rem;
-  min-height: 100%;
-}
-
-.container {
-  width: 80%;
-}
-
-.theme-background,
-.theme-background:hover,
-.theme-background:focus {
-  background-color: #3C849C;
-}
-
-.donation-banner {
-  padding: 10px 20px;
-  background-color: #DCF2FA;
-  color: #3C849C;
-}
-
-.footer {
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  padding: 1rem;
-  color: #ffffff;
-}
